# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AndMI API/Open API System",
        "description": "Andijon Mashinasozlik Instituti Ochiq API tizimi",
        "contact": {
            "email": "andmiatmdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/news_rus": {
            "get": {
                "tags": [
                    "Новости"
                ],
                "summary": "Получение списка всех новостей на русском",
                "operationId": "133510652b8efd50e56237c75d3dd761",
                "responses": {
                    "200": {
                        "description": "Список новостей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Заголовок новости"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "slug-novosti"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Содержание новости..."
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "file_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "views": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "published": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Новости"
                ],
                "summary": "Создание новости на русском",
                "operationId": "30c81b643d3e9428bc286b9b72b38866",
                "requestBody": {
                    "description": "Данные для создания новости",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "content",
                                    "likes",
                                    "file_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Заголовок новости"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "slug-novosti"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Содержание новости..."
                                    },
                                    "likes": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "file_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "views": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Новость успешно создана"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/news_rus/{id}": {
            "get": {
                "tags": [
                    "Новости"
                ],
                "summary": "Получение новости на русском",
                "operationId": "5b67cebc59fb0e8b1982ee0c664c5407",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новости",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Новость найдена"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Новость не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Новости"
                ],
                "summary": "Обновление новости на русском",
                "operationId": "d974f130a1e65e1aa2cee72f00e7f1f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новости",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления новости",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "content",
                                    "likes",
                                    "file_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Обновленный заголовок новости"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "updated-slug-novosti"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Обновленное содержание новости..."
                                    },
                                    "likes": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "file_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "views": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Новость успешно обновлена"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Новость не найдена"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Новости"
                ],
                "summary": "Удаление новости на русском",
                "operationId": "dddd12825b07f8f9a45be3313b8391dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новости",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Новость успешно удалена"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Новость не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token orqali autentifikatsiyadan o'tish",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}